Refactoring for PlayerCog
More refactoring for Player cog.
Start of implementation of priorityqueue instead of normal queue. This is due to the play command having priority over the yt command. So if a youtube video is playing and a .play command is invoked the youtube video should be paused, mp3 file should be played, then start playing the youtube video again.