Casino cog begun development
Casino cog has been created! Along with 37 gifs for each of the different spins in the roulette game. Currently the Casino only has a very simple roulette game. Only possible for one person to play on a single spin. No currency implemented yet so you just spin for the fun.
Configuration cog has been renamed as it needs to be loaded as the first cog since it is used in all the other cogs.